refactor(backend): extract document preview payloads
- Move uploaded-document lesson and course preview payload builders into a focused module.\n- Keep direct tool-round compatibility exports for existing private tests and callers.\n- Update architecture and cleanup docs to reflect the new runtime boundary.
